+++ src/server/graphql/blockchain/fetchers/transactions.ts
@@ -12,11 +12,14 @@ import {
     getNodeSelectionSetForConnection,
     isDefined,
     prepareChainOrderFilter,
+    prepareNonChainOrderPaginationFilter,
     processPaginatedQueryResult,
     processPaginationArgs,
 } from "../helpers"
 import {
     BlockchainAccountQueryTransactionsArgs,
+    BlockchainAccountQueryTransactions_By_LtArgs,
+    BlockchainQueryTransactions_By_In_MsgArgs,
     BlockchainQueryTransactionsArgs,
     BlockchainTransaction,
     BlockchainTransactionsConnection,
@@ -77,6 +80,56 @@ export async function resolve_transaction(
     )
 
     return queryResult[0]
+}
+
+export async function resolve_transactions_by_in_msg(
+    args: BlockchainQueryTransactions_By_In_MsgArgs,
+    context: QRequestContext,
+    info: GraphQLResolveInfo,
+    traceSpan: QTraceSpan,
+) {
+    const maxJoinDepth = 2
+
+    const useArchive = useTransactionsArchive(args.archive, context)
+    const { selectionSet, requireBocParsing } = upgradeSelectionForBocParsing(
+        useArchive,
+        info.fieldNodes[0].selectionSet,
+        transactionArchiveFields,
+    )
+
+    const returnExpression = config.transactions.buildReturnExpression(
+        selectionSet,
+        context,
+        maxJoinDepth,
+        "doc",
+    )
+
+    // query
+    const params = new QParams({
+        stringifyKeyInAqlComparison:
+            context.services.config.queries.filter.stringifyKeyInAqlComparison,
+    })
+    const query =
+        "FOR doc IN transactions " +
+        `FILTER doc.in_msg == @${params.add(args.msg_hash)} ` +
+        "SORT doc.lt ASC LIMIT 50 " +
+        `RETURN ${returnExpression}`
+    const queryResult = await parseTransactionBocsIfRequired(
+        requireBocParsing,
+        context,
+        (await context.services.data.query(
+            required(context.services.data.transactions.provider),
+            {
+                text: query,
+                vars: params.values,
+                orderBy: [],
+                request: context,
+                traceSpan,
+                archive: useArchive,
+            },
+        )) as BlockchainTransaction[],
+    )
+    return queryResult
 }
 
 export async function resolve_blockchain_transactions(
@@ -164,6 +217,7 @@ export async function resolve_blockchain_transactions(
                 context,
                 traceSpan,
                 maxJoinDepth,
+                useArchive,
             )
         },
     )) as BlockchainTransactionsConnection
@@ -254,6 +308,95 @@ export async function resolve_account_transactions(
                 context,
                 traceSpan,
                 maxJoinDepth,
+                useArchive,
+            )
+        },
+    )) as BlockchainTransactionsConnection
+}
+
+export async function resolve_account_transactions_by_lt(
+    account_address: string,
+    args: BlockchainAccountQueryTransactions_By_LtArgs,
+    context: QRequestContext,
+    info: GraphQLResolveInfo,
+    traceSpan: QTraceSpan,
+) {
+    const maxJoinDepth = 2
+    // filters
+    const filters: string[] = []
+    const params = new QParams({
+        stringifyKeyInAqlComparison:
+            context.services.config.queries.filter.stringifyKeyInAqlComparison,
+    })
+
+    await prepareNonChainOrderPaginationFilter(
+        args,
+        params,
+        filters,
+        context,
+        "lt",
+    )
+    filters.push(`doc.account_addr == @${params.add(account_address)}`)
+
+    const { direction, limit } = processPaginationArgs(args)
+
+    const useArchive = useTransactionsArchive(args.archive, context)
+    const { selectionSet, requireBocParsing } = upgradeSelectionForBocParsing(
+        useArchive,
+        getNodeSelectionSetForConnection(info),
+        transactionArchiveFields,
+    )
+
+    const returnExpression = config.transactions.buildReturnExpression(
+        selectionSet,
+        context,
+        maxJoinDepth,
+        "doc",
+        ["lt"],
+    )
+
+    // query
+    const query = `
+        FOR doc IN transactions
+        FILTER ${filters.join(" AND ")}
+        SORT doc.lt ${direction == Direction.Backward ? "DESC" : "ASC"}
+        LIMIT ${limit}
+        RETURN ${returnExpression}
+    `
+    const queryResult = await parseTransactionBocsIfRequired(
+        requireBocParsing,
+        context,
+        (await context.services.data.query(
+            required(context.services.data.transactions.provider),
+            {
+                text: query,
+                vars: params.values,
+                orderBy: [
+                    {
+                        path: "lt",
+                        direction: "ASC",
+                    },
+                ],
+                request: context,
+                traceSpan,
+                archive: useArchive,
+            },
+        )) as BlockchainTransaction[],
+    )
+
+    return (await processPaginatedQueryResult(
+        queryResult,
+        limit,
+        direction,
+        "lt",
+        async r => {
+            await config.transactions.fetchJoins(
+                r,
+                selectionSet,
+                context,
+                traceSpan,
+                maxJoinDepth,
+                useArchive,
             )
         },
     )) as BlockchainTransactionsConnection